Test/requirements.txt
behave
websocket-client
sseclient
requests
assertpy
chromedriver-binary